+++ src/main/scala/io/viash/ViashTest.scala
@@ -61,7 +61,7 @@ object ViashTest {
     platform: Platform,
     keepFiles: Option[Boolean] = None,
     quiet: Boolean = false,
-    setupStrategy: String = "cachedbuild",
+    setupStrategy: Option[String] = None,
     tempVersion: Option[String] = Some("test"),
     verbosityLevel: Int = 6,
     parentTempPath: Option[Path] = None, 
@@ -90,7 +90,7 @@ object ViashTest {
       platform = platform,
       dir = dir,
       verbose = !quiet,
-      setupStrategy = setupStrategy,
+      setupStrategy = setupStrategy.getOrElse("cachedbuild"),
       verbosityLevel = verbosityLevel,
       cpus = cpus,
       memory = memory