+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Viash 0.x.x (yyyy-MM-dd): TODO Add title
 
-
 ## BREAKING CHANGES
 
 * `viash export cli_schema`: Added `--format yaml/json` argument, default format is now a YAML (PR #448).
@@ -11,6 +10,8 @@
 * `viash export json_schema`: Export a json schema derived from the class reflections and annotations already used by the `config_schema` (PR #446).
 
 * `viash export config_schema`: Output `default` values of member fields (PR #446).
+
+* `viash test` and `viash ns test`: add `--setup` argument to determine the docker build strategy before a component is tested (PR #451).
 
 ## MINOR CHANGES